requirements.txt
SQLAlchemy==2.0.5.post1
PyYAML==6.0
starlette==0.25.0
uvicorn==0.20.0
fastapi==0.93.0
pydantic==1.10.6